// src/app/api/nswag-client.ts
export interface RegisterRequest {
    name?: string | undefined;
    email?: string | undefined;
    password?: string | undefined;
}