// src/app/api/model/organizationCreateDto.ts
/**
 * SmartLedger API
 *
 * Contact: support@smartledger.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrganizationCreateDto { 
    name?: string | null;
    industry?: string | null;
    gstNumber?: string | null;
    country?: string | null;
    plan?: string | null;
}